requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
mangum==0.17.0
whitenoise==6.6.0
aws-lambda-powertools==2.30.2 